+++ src/main/java/org/jboss/elemento/svg/SVGElement.java
@@ -25,7 +25,8 @@ import jsinterop.base.JsPropertyMap;
 /**
  * All of the SVG DOM interfaces that correspond directly to elements in the SVG language derive from the SVGElement interface.
  *
- * @see <a href="https://developer.mozilla.org/en-US/docs/Web/API/SVGElement">https://developer.mozilla.org/en-US/docs/Web/API/SVGElement</a>
+ * @see <a href=
+ *      "https://developer.mozilla.org/en-US/docs/Web/API/SVGElement">https://developer.mozilla.org/en-US/docs/Web/API/SVGElement</a>
  */
 @JsType(isNative = true, namespace = JsPackage.GLOBAL)
 public class SVGElement extends Element {