+++ src/main/java/org/jboss/elemento/svg/SVGGElement.java
@@ -21,8 +21,8 @@ import jsinterop.annotations.JsType;
 /**
  * The SVGGElement interface corresponds to the {@code <g>} element.
  *
- * @see <a
- * href="https://developer.mozilla.org/en-US/docs/Web/API/SVGGElement">https://developer.mozilla.org/en-US/docs/Web/API/SVGGElement</a>
+ * @see <a href=
+ *      "https://developer.mozilla.org/en-US/docs/Web/API/SVGGElement">https://developer.mozilla.org/en-US/docs/Web/API/SVGGElement</a>
  */
 @JsType(isNative = true, namespace = JsPackage.GLOBAL)
 public class SVGGElement extends SVGElement {