+++ src/main/java/org/jboss/elemento/svg/SVGLineElement.java
@@ -19,11 +19,11 @@ import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 
 /**
- * The SVGLineElement interface provides access to the properties of {@code <line>} elements, as well as methods to
- * manipulate them.
+ * The SVGLineElement interface provides access to the properties of {@code <line>} elements, as well as methods to manipulate
+ * them.
  *
- * @see <a
- * href="https://developer.mozilla.org/en-US/docs/Web/API/SVGLineElement">https://developer.mozilla.org/en-US/docs/Web/API/SVGLineElement</a>
+ * @see <a href=
+ *      "https://developer.mozilla.org/en-US/docs/Web/API/SVGLineElement">https://developer.mozilla.org/en-US/docs/Web/API/SVGLineElement</a>
  */
 @JsType(isNative = true, namespace = JsPackage.GLOBAL)
 public class SVGLineElement extends SVGElement {