+++ src/main/java/org/jboss/elemento/svg/SVGPathElement.java
@@ -21,8 +21,8 @@ import jsinterop.annotations.JsType;
 /**
  * The SVGPathElement interface corresponds to the {@code <path>} element.
  *
- * @see <a
- * href="https://developer.mozilla.org/en-US/docs/Web/API/SVGPathElement">https://developer.mozilla.org/en-US/docs/Web/API/SVGPathElement</a>
+ * @see <a href=
+ *      "https://developer.mozilla.org/en-US/docs/Web/API/SVGPathElement">https://developer.mozilla.org/en-US/docs/Web/API/SVGPathElement</a>
  */
 @JsType(isNative = true, namespace = JsPackage.GLOBAL)
 public class SVGPathElement extends SVGElement {