+++ src/main/java/org/jboss/elemento/svg/SVGRectElement.java
@@ -19,11 +19,11 @@ import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 
 /**
- * The SVGRectElement interface provides access to the properties of {@code <rect>} elements, as well as methods to
- * manipulate them.
+ * The SVGRectElement interface provides access to the properties of {@code <rect>} elements, as well as methods to manipulate
+ * them.
  *
- * @see <a
- * href="https://developer.mozilla.org/en-US/docs/Web/API/SVGRectElement">https://developer.mozilla.org/en-US/docs/Web/API/SVGRectElement</a>
+ * @see <a href=
+ *      "https://developer.mozilla.org/en-US/docs/Web/API/SVGRectElement">https://developer.mozilla.org/en-US/docs/Web/API/SVGRectElement</a>
  */
 @JsType(isNative = true, namespace = JsPackage.GLOBAL)
 public class SVGRectElement extends SVGElement {